requirements.txt
conceptnet_lite==0.1.27
nltk==3.8.1
pandas==2.0.1
PyYAML==6.0.1
neo4j==5.12.0
openpyxl==3.1.2